Déplacement des méthodes dans plusieurs fichiers sources
